refactor(wallet-sdk): accounts feature constructs its own repository
The web builds its account repository from inputs it already holds
(db client, session-key getters via /temporary), the same shape the
receive feature uses; the SDK-side repository accessor and its
/temporary export are removed.
