+++ packages/wallet-sdk/domain/sdk/sdk.test.ts
@@ -1,11 +1,7 @@
 import { describe, expect, it } from 'bun:test';
 import type { AuthKeyValueStore, SdkConfig } from '.';
 import { nullLogger } from '../../lib/logger';
-import {
-  AgicashSdk,
-  getInternalAccountRepository,
-  getInternalSessionKeys,
-} from './sdk';
+import { AgicashSdk, getInternalSessionKeys } from './sdk';
 
 const createMemoryStore = (): AuthKeyValueStore => {
   const data = new Map<string, string>();
@@ -45,34 +41,6 @@ describe('AgicashSdk.create', () => {
 
     const next = AgicashSdk.create(createConfig());
     await next.dispose();
-  });
-});
-
-describe('getInternalAccountRepository', () => {
-  it('throws when no instance is live', () => {
-    expect(() => getInternalAccountRepository()).toThrow(
-      'No current AgicashSdk instance',
-    );
-  });
-
-  it('resolves through the current instance and clears on dispose', async () => {
-    const sdk = AgicashSdk.create(createConfig());
-
-    // While live the guard passes and the accessor returns the repository
-    // promise. Its key derivation is exercised in the accounts-api tests; here
-    // we only assert the bridge is wired, so the derivation is left to settle.
-    const pending = getInternalAccountRepository();
-    expect(pending).toBeInstanceOf(Promise);
-    pending.catch(() => {
-      // The derivation reaches Open Secret, which this env cannot; the bridge
-      // wiring is all we assert here, so let the rejection settle unobserved.
-    });
-
-    await sdk.dispose();
-
-    expect(() => getInternalAccountRepository()).toThrow(
-      'No current AgicashSdk instance',
-    );
   });
 });
